+++ src/components/header/NavigationPanel/NavigationPanel.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const NavigationPanel = () => {
+  return (
+    <div>NavigationPanel</div>
+  )
+}
+
+export default NavigationPanel
