+++ src/components/header/index.tsx
@@ -1,0 +1,2 @@
+export { default as DropDownBox } from "./DropDownBox/DropDownBox";
+export { default as HeaderInput } from "./HeaderInput/HeaderInput";
